Integratehttp YOU HAVE TO UPDATE DEFINES ON WIFI.H FOR HOTSPOT AND .url HTTP STRUCT EVERYTIME DEVICES CONNECT TO NEW NETWORKK

+++ microSrc/src/dataTypes/plantData.h
@@ -9,7 +9,7 @@
 struct plantData
 {
     uint16_t lightData;
-    uint8_t waterData; // Rough % of wetness
+    uint16_t waterData; // Rough % of wetness
     uint16_t nitrogenData;
     uint16_t phosphorusData;
     uint16_t potassiumData;
